+++ torch/csrc/jit/passes/onnx/constant_fold.cpp
@@ -472,7 +472,7 @@ c10::optional<at::Tensor> runTorchBackendForOnnx(
     // If indices input for onnx::Gather has a value less than 0,
     // It needs to be adjusted (+= dim value) for aten op
     auto less_mask = at::lt(indices, 0);
-    auto indices_corr = at::add(indices, inputTensorValues[0].sizes()[axis]);
+    auto indices_corr = at::add(indices, (int64_t)inputTensorValues[0].sizes()[axis]);
     auto indices_masked = at::where(less_mask, indices_corr, indices);
     updated_val = at::index_select(inputTensorValues[0], axis, indices_masked);
     // If rank of indices is 0, rank of output tensor should be
